Removiendo campo de API key (manejo de única API key)
