+++ .env.example
@@ -1,0 +1,1 @@
+GOOGLE_GENERATIVE_AI_API_KEY=<get key at https://aistudio.google.com/app/>
